feat: поддержка нескольких очередей

+++ src/JobQueue/JobQueue.Core/IJobQueue.cs
@@ -2,7 +2,7 @@
 
 public interface IJobQueue
 {
-    public bool TryEnqueue(int key, byte[] payload);
-    public bool TryDequeue(out int key, out byte[] payload);
+    public bool TryEnqueue(long key, byte[] payload);
+    public bool TryDequeue(out long key, out byte[] payload);
     public int Count { get; }
 }
